extract common code into base_import, recodes, and canon files
