+++ reports/utils/recodes.py
@@ -1,0 +1,140 @@
+RECODES = {
+    # country + region
+    'Pacific': 'Pacific Islands',
+    'Congo, Rep (Brazzaville)': 'Congo',
+    'Congo, Dem Rep': 'Congo (the Democratic Republic of the)',
+    'Bosnia & Herzegovina': 'Bosnia and Herzegovina',
+    'St Lucia': 'Saint Lucia',
+    'St. Vincent and The Grenadines': 'Saint Vincent and the Grenadines',
+    'Trinidad & Tobago': 'Trinidad and Tobago',
+    # women topics
+    "Arts, entertainment, leisure, cinema, theatre, books, dance": "Arts, entertainment, leisure, cinema, books, dance",
+    "Beauty contests, models, fashion, beauty aids, cosmetic surgery": "Beauty contests, models, fashion, cosmetic surgery",
+    "Birth control, fertility, sterilisation, amniocentesis, termination of pregnancy": "Birth control, fertility, sterilization, termination...",
+    "Celebrity news, births, marriages, deaths, obituaries, famous people, royalty": "Celebrity news, births, marriages, royalty, etc.",
+    "Changing gender relations, roles and relationships of women and men inside and outside the home": "Changing gender relations (outside the home)",
+    "Child abuse, sexual violence against children, trafficking, neglect.": "Child abuse, sexual violence against children, neglect",
+    "Consumer issues, consumer protection, regulation, prices, consumer fraud": "Consumer issues, consumer protection, fraud...",
+    "Development issues, sustainability, community development": "Other development issues, sustainability, etc.",
+    "Disaster, accident, famine, earthquake, flood, hurricane, plane crash, car crash": "Disaster, accident, famine, flood, plane crash, etc.",
+    "Economic crisis, state bailouts of companies, company takeovers and mergers": "Economic crisis, state bailouts of companies, company takeovers and mergers, etc.",
+    "Education, child care, nurseries, pre-school to university, adult education, literacy": "Education, childcare, nursery, university, literacy",
+    "Environment, nature,   pollution, global warming, ecology, tourism": "Environment, pollution, tourism",
+    "Family law, family codes, property law, inheritance law and rights": "Family law, family codes, property law, inheritance...",
+    "Family relations, inter-generational conflict, single parents": "Family relations, inter-generational conflict, parents",
+    "Foreign/international politics, relations with other countries, negotiations, treaties, UN peacekeeping": "Foreign/international politics, UN, peacekeeping",
+    "Gender-based violence, feminicide, harassment, domestic violence, rape, trafficking, genital mutilation": "Gender violence based on culture, family, inter-personal relations, feminicide, harassment, rape, sexual assault, trafficking, FGM...",
+    "HIV and AIDS, incidence, policy, treatment, people affected": "HIV and AIDS, policy, treatment, etc",
+    "Human rights, women's rights, children's rights, gay & lesbian rights, rights of minorities ..": "Human rights, womens rights, rights of sexual minorities, rights of religious minorities, etc.",
+    "Legal system, judicial system, legislation (apart from family, property & inheritance law)": "Legal system, judiciary, legislation apart from family",
+    "Media, including new media (computers, internet), portrayal of women and/or men, pornography": "Media, (including internet), portrayal of women/men",
+    "Medicine, health, hygiene, safety, disability, medical research, funding (apart from HIV-AIDS)": "Medicine, health, hygiene, safety, (not EBOLA or HIV/AIDS)",
+    "Migration, refugees, asylum seekers, ethnic conflict, integration, racism, xenophobia": "Migration, refugees, xenophobia, ethnic conflict...",
+    "National defence, military spending, military training, military parades, internal security": "National defence, military spending, internal security, etc.",
+    "Non-violent crime, bribery, theft, drug-dealing, corruption, (including political corruption/malpractice)": "Non-violent crime, bribery, theft, drugs, corruption",
+    "Other epidemics, viruses, contagions, Influenza, BSE, SARS": "Other epidemics, viruses, contagions, Influenza, BSE, SARS",
+    "Other labour issues, strikes, trade unions, negotiations, other employment and unemployment": "Other labour issues (strikes, trade unions, etc.)",
+    "Other stories on celebrities, arts, media (specify the subject in 'Comments' section of coding sheet)": "Other stories on politics (specify in comments)",
+    "Other stories on crime and violence (specify the subject in 'Comments' section of coding sheet)": "Other stories on science (specify in comments)",
+    "Other stories on politics and government (specify the subject in 'Comments' section of coding sheet)": "Other stories on politics (specify in comments)",
+    "Other stories on science or health (specify the subject in 'Comments' section of coding sheet)": "Other stories on science (specify in comments)",
+    "Other stories on social or legal issues (specify the subject in 'Comments' section of coding sheet)": "Other stories on social/legal (specify in comments)",
+    "Other stories on the economy (specify the subject in 'Comments' section of coding sheet)": "Other stories on economy (specify in comments)",
+    "Peace, negotiations, treaties(local, regional, national),": "Peace, negotiations, treaties",
+    "Poverty, housing, social welfare, aid to those in need": "Poverty, housing, social welfare, aid, etc.",
+    "Religion, culture, tradition, controversies, teachings, celebrations, practices": "Religion, culture, tradition, controversies...",
+    "Riots, demonstrations, public disorder": "Riots, demonstrations, public disorder, etc.",
+    "Rural economy, agriculture, farming practices, agricultural policy, land rights": "Rural economy, agriculture, farming, land rights",
+    "Science, technology, research, funding, discoveries, developments": "Science, technology, research, discoveries...",
+    "Sports, events, players, facilities, training, policies, funding": "Sports, events, players, facilities, training, funding",
+    "Transport, traffic, roads": "Transport, traffic, roads...",
+    "Violent crime, murder, abduction, kidnapping, assault, drug-related violence": "Violent crime, murder, abduction, assault, etc.",
+    "War, civil war, terrorism, state-based violence": "War, civil war, terrorism, other state-based violence",
+    "Women's movement, activism, events, demonstrations, gender equality advocacy": "Womens movement, activism, demonstrations, etc",
+    "Other domestic politics/government (local, regional, national), elections, speeches, the political process": "Other domestic politics, government, etc.",
+    "Global partnerships (international trade and finance systems, e.g. WTO, IMF, World Bank, debt)": "Global partnerships",
+    # occupations
+    "Agriculture, mining, fishing, forestry worker": "Agriculture, mining, fishing, forestry",
+    "Celebrity, artist, actor, writer, singer, radio or television personality": "Celebrity, artist, actor, writer, singer, TV personality",
+    "Government employee, public servant, bureaucrat, diplomat, intelligence officer": "Government employee, public servant, etc.",
+    "Homemaker, parent, either female or male. Code this only if no other occupation is given, e.g. a doctor who is also described as a mother is coded 6.": "Homemaker, parent (male or female)) only if no other occupation is given e.g. doctor/mother=code 6",
+    "Lawyer, judge, magistrate, legal advocate, legal expert, legal clerk": "Lawyer, judge, magistrate, legal advocate, etc.",
+    "Media professional, journalist, video or film-maker, theatre director ...": "Media professional, journalist, film-maker, etc.",
+    "Office or service worker, non-management worker in office, store, restaurant, catering": "Office or service worker, non-management worker",
+    "Police, military, para-military group, militia, prison officer, security officer, fire officer": "Police, military, para-military, militia, fire officer",
+    "Religious figure, priest, monk, rabbi, mullah, nun": "Religious figure, priest, monk, rabbi, mullah, nun",
+    "Royalty, ruling monarch, deposed monarch, any member of royal family": "Royalty, monarch, deposed monarch, etc.",
+    "Science or technology professional, engineer, technician, computer specialist": "Science/ technology professional, engineer, etc.",
+    "Sportsperson, athlete, player, coach, referee": "Sportsperson, athlete, player, coach, referee",
+    "Student, pupil, schoolchild": "Student, pupil, schoolchild",
+    "Tradesperson, artisan, labourer, truck driver, construction, factory, domestic worker": "Tradesperson, artisan, labourer, truck driver, etc.",
+    "Government official, politician, president, government minister, political leader, political party staff, spokesperson": "Government, politician, minister, spokesperson...",
+    "Business person, executive, manager, entrepreneur, economist, financial expert, stock broker": "Business person, exec, manager, stock broker...",
+    "Child, young person (up to 18 years). Code this only if no other occupation/position is given, e.g. a schoolchild is coded 19; a child labourer is coded 12.": "Child, young person no other occupation given",
+    "Villager or resident engaged in unspecified occupation. Code this only if no other occupation is given, e.g. a teacher who is also described as a villager is coded 5.": "Villager or resident no other occupation given",
+    "Criminal, suspect. Code this only if no other occupation is given, e.g, a lawyer suspected of committing a crime is coded 9; a former politician who has committed a crime is coded 2.": "Criminal, suspect no other occupation given",
+    "Unemployed. Code this only if no other occupation is given, e.g. an unemployed actor is coded 17; an unemployed person who commits a crime is coded 24.": "Unemployed no other occupation given",
+    "Sex worker, prostitute": "Sex worker",
+    "Other. Use only as a last resort (specify the occupation/position in 'Comments' section of coding sheet)": "Other only as last resort & explain",
+    "Retired person, pensioner. Code this only if no other occupation is given, e.g. a retired police officer is coded 4; a retired politician is coded 2.": "Retired person, pensioner no other occupation given",
+    "Activist or worker in civil society organisation, non-governmental organisation, trade union, human rights, consumer issues, environment, aid agency, peasant leader, United Nations": "Activist or worker in civil society org., NGO, trade union",
+    # function
+    "Subject: the story is about this person, or about something the person has done, said etc.": "Subject",
+    "Spokesperson: the person represents, or speaks on behalf of another person, a group or an organisation": "Spokesperson",
+    "Expert or commentator: the person provides additional information, opinion or comment, based on specialist knowledge or expertise": "Expert or commentator",
+    "Personal experience: the person provides opinion or comment, based on individual personal experience; the opinion is not necessarily meant to reflect the views of a wider group": "Personal Experience",
+    "Eye witness: the person gives testimony or comment, based on direct observation (e.g. being present at an event)": "Eye Witness",
+    "Popular opinion: the person's opinion is assumed to reflect that of the 'ordinary citizen' (e.g., in a street interview, vox populi etc); it is implied that the person's point of view is shared by a wider group of people.": "Popular Opinion",
+    "Other. Use only as a last resort (describe the function in 'Comments' section of coding sheet).": "Other",
+    "12 years or under": "12 and under",
+    # sex
+    "%F": "Female",
+    "%M": "Male",
+    'Female  %F': 'Female',
+    "Female %": 'Female',
+    "Male %F": "Male",
+    "Male %f": "Male",
+    "Male %M": "Male",
+    "Female %": "Female",
+    'Female  %F"': "Female",
+    "Male %": "Male",
+    "Other: transgender, transsexual": "Other (transgender, etc.)",
+    # survivor_of
+    "Survivor of an accident, natural disaster, poverty, disease, illness": "Survivor of an accident, natural disaster, poverty",
+    "Survivor of domestic violence (by husband/wife/partner/other family member), psychological violence, physical assault, marital rape, murder": "Survivor of domestic violence, rape, murder, etc.",
+    "Survivor of domestic violence (by husband/wife/partner/other family member), psychological violence, physical assault, marital rape, murder": "Survivor of domestic violence, rape, murder, etc.",
+    "Survivor of non-domestic sexual violence or abuse, sexual harassment, rape, trafficking": "Survivor of non-domestic sexual violence, rape, assault, etc. (sexual violence only)",
+    "Survivor of other crime, robbery, assault, murder": "Survivor of other non-domestic crime, robbery, etc.",
+    "Survivor of violation based on religion, tradition, cultural belief, genital mutilation, bride-burning": "Survivor of violation based on religion, tradition...",
+    "Survivor of war, terrorism, vigilantism, state-based violence": "Survivor of war, terrorism, vigilantism, state violence...",
+    "Survivor of discrimination based on gender, race, ethnicity, age, religion": "Survivor of discrimination based on gender, race, ethnicity, age, religion, ability, etc.",
+    "Other survivor: describe in 'Comments' section of coding sheet": "Other survivor (specify in comments)",
+    # victim_of
+    "Victim of an accident, natural disaster, poverty, disease, illness": "Victim of an accident, natural disaster, poverty",
+    "Victim of domestic violence (by husband/wife/partner/other family member), psychological violence, physical assault, marital rape, murder": "Victim of domestic violence, rape, murder, etc.",
+    "Victim of non-domestic sexual violence or abuse, sexual harassment, rape, trafficking": "Victim of non-domestic sexual violence, rape, assault, etc (sexual violence only)",
+    "Victim of other crime, robbery, assault, murder": "Victim of other non-domestic crime, robbery, etc.",
+    "Victim of violation based on religion, tradition, cultural belief, genital mutilation, bride-burning": "Victim of violation based on religion, tradition...",
+    "Victim of war, terrorism, vigilantism, state-based violence": "Victim of war, terrorism, vigilantism, state violence...",
+    "Victim of discrimination based on gender, race, ethnicity, age, religion, ability": "Victim of discrimination based on gender, race, ethnicity, age, religion, ability, etc",
+    "Other victim: describe in 'Comments' section of coding sheet": "Other victim (specify in comments)",
+    # yes/no,
+    "No, women are not central": "No",
+    "Yes, women are central": "Yes",
+    # region weirdness
+    'Africa Female  %F': 'Africa',
+    'Asia Female  %F': 'Asia',
+    'Caribbean Female  %F': 'Caribbean',
+    'Europe Female  %F': 'Europe',
+    'Latin America Female  %F': 'Latin America',
+    'Middle East Female  %F': 'Middle East',
+    'North America Female  %F': 'North America',
+    'Pacific Female  %F': 'Pacific Islands',
+    "Do not know: (e.g., the person may not be seen clearly)": "Do not know",
+    # countries
+    "United Kingdom (incl Northern Ireland, Scotland, Wales)": "United Kingdom - England, Ireland, Scotland and Wales",
+    "Belgium": "Belgium - French and Flemish",
+}
+
+def recode(v):
+    return RECODES.get(v, v)
